`patchEnvelopes` use `PatchEnvelopeInfo` between zedrouter and zedagent
This is first part of implementing external patch envelopes.

Patch envelopes are nothing more but a data exposed to Edge Applications
by EVE in runtime. This is done by metadata server which is part of zedrouter.
There are two types of patch envelopes: inline and external.

Inline patch envelopes are small objects (10Kbytes max as of now) which
are part of `EdgeDevConfig`.

External patch envelopes are volumes, downloaded by EVE and exposed to
Edge Applications via metadata server. Volumes are handled by other `volumemgr`
service which means that with external patch envelopes we need to get updates
not only from `zedagent` but also from `volumemgr` and those updates are not in
sync. So in order to handle that we need to

1) Separate information sent by `zedagent` to what we actually expose in `zedrouter`
2) Implement structure which will handle async updates from `volumemgr` and `zedagent`
3) Switch metdata server to using structure implemented in 2)

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>