`persistcache` add KeyNotFoundError and respecive handle in Get
In case when user tries to get key which is not in PersistCache
KeyNotFoundError will be returned. That allows for graceful handling
of situations when key is not present.

Alternative to this approach would be to add function Has() which
checks if Key present in PersistCache, however this library tries to
keep deep interfraces with minimum functions.

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>